Add ArchLinux pkgbuild
The pkgbuild applies two patches which solve some issues with the makefile install process on ArchLinux
The first patch removes the call to ldconfig in the makefile, as we will apply it in the .install file later on.
The second patch changes the install prefix to /usr from /usr/local, which is the standard location on Arch systems

The .gitignore excludes some build files which otherwise cause annoyance.