Remove the argument parsing, expect FPS to be first argument
This patch removes the loop which parses arguments looking for an FPS=/fps= argument.
Other programs may accept fps=* as a valid argument, and libstrangle would eat the
argument with the loop.

By removing the loop, we are able to change two things.
One is that it would allow these hypothetical programs to run correctly.
Two is that it makes the libstrangle input more restrcitve, the FPS must be the first argument, and
  all following arguments are expected to be the program to strangle
Three is that it allows the script to run using plain /bin/sh instead of relying on bash (and bash arrays)